Pixelated Vanilla-Style Lava
This is probably a bit messy, but it's the result of tweaking the lava a bunch until I got something that fits with the vanilla lava texture. I'm not sure if you want to use this as the main lava shader, or leave it as an option, but I want to share it somewhere in case people want to use the modification.